+++ pydbml/_classes/reference.py
@@ -19,7 +19,7 @@ class Reference(SQLObject, DBMLObject):
     and its `sql` property contains the ALTER TABLE clause.
     '''
     required_attributes = ('type', 'col1', 'col2')
-    dont_compare_fields = ('database', '_inline')
+    _eq_skip_fields = ('database', '_inline')
 
     def __init__(self,
                  type: Literal['>', '<', '-', '<>'],
@@ -62,7 +62,7 @@ class Reference(SQLObject, DBMLObject):
             name=f'{self.table1.name}_{self.table2.name}',
             schema=self.table1.schema,
             columns=(
-                Column(name=f'{c.table.name}_{c.name}', type=c.type, not_null=True, pk=True)  # type: ignore
+                Column(name=f'{c.table.name}_{c.name}', type=c.type, not_null=True, pk=True)  # type: ignore[union-attr]
                 for c in chain(self.col1, self.col2)
             ),
             abstract=True
@@ -99,18 +99,18 @@ class Reference(SQLObject, DBMLObject):
         >>> c1 = Column('c1', 'int')
         >>> c2 = Column('c2', 'int')
         >>> print(Reference('>', col1=c1, col2=c2))
-        Reference([c1] > [c2]
+        Reference([c1] > [c2])
         >>> c12 = Column('c12', 'int')
         >>> c22 = Column('c22', 'int')
         >>> print(Reference('<', col1=[c1, c12], col2=(c2, c22)))
-        Reference([c1, c12] < [c2, c22]
+        Reference([c1, c12] < [c2, c22])
         '''
 
         col1 = ', '.join(f'{c.name}' for c in self.col1)
         col2 = ', '.join(f'{c.name}' for c in self.col2)
-        return f"Reference([{col1}] {self.type} [{col2}]"
+        return f"Reference([{col1}] {self.type} [{col2}])"
 
-    def _validate(self):
+    def _validate(self) -> None:
         table1 = self.col1[0].table
         if any(c.table != table1 for c in self.col1):
             raise DBMLError('Columns in col1 are from different tables')