+++ pydbml/constants.py
@@ -2,3 +2,5 @@ ONE_TO_MANY = '<'
 MANY_TO_ONE = '>'
 ONE_TO_ONE = '-'
 MANY_TO_MANY = '<>'
+
+DEFAULT_SCHEMA = 'public'