+++ pydbml/renderer/dbml/default/reference.py
@@ -6,20 +6,21 @@ from pydbml.classes import Reference, Column
 from pydbml.exceptions import TableNotFoundError, DBMLError
 from pydbml.renderer.dbml.default.renderer import DefaultDBMLRenderer
 from pydbml.renderer.dbml.default.utils import comment_to_dbml
-from .table import get_full_name_for_dbml
+from pydbml.renderer.utils import get_full_name
 
 
 def validate_for_dbml(model: Reference):
     for col in chain(model.col1, model.col2):
         if col.table is None:
-            raise TableNotFoundError(f'Table on {col} is not set')
+            raise TableNotFoundError(f'Table for column {col!r} is not set')
 
 
 def render_inline_reference(model: Reference) -> str:
     # settings are ignored for inline ref
     if len(model.col2) > 1:
         raise DBMLError('Cannot render DBML: composite ref cannot be inline')
-    table_name = get_full_name_for_dbml(model.col2[0].table)
+    assert model.col2[0].table is not None  # guaranteed by validate_for_dbml
+    table_name = get_full_name(model.col2[0].table)
     return f'ref: {model.type} {table_name}."{model.col2[0].name}"'
 
 
@@ -43,16 +44,17 @@ def render_options(model: Reference) -> str:
 
 
 def render_not_inline_reference(model: Reference) -> str:
+    assert model.table1 is not None and model.table2 is not None  # guaranteed by validate_for_dbml
     result = comment_to_dbml(model.comment) if model.comment else ''
     result += 'Ref'
     if model.name:
         result += f' {model.name}'
 
     result += (
-        ' {\n    '  # type: ignore
-        f'{get_full_name_for_dbml(model.table1)}.{render_col(model.col1)} '
+        ' {\n    '
+        f'{get_full_name(model.table1)}.{render_col(model.col1)} '
         f'{model.type} '
-        f'{get_full_name_for_dbml(model.table2)}.{render_col(model.col2)}'
+        f'{get_full_name(model.table2)}.{render_col(model.col2)}'
         f'{render_options(model)}'
         '\n}'
     )