+++ pydbml/tools.py
@@ -9,7 +9,7 @@ def comment(val: str, comb: str) -> str:
     return '\n'.join(f'{comb} {cl}' for cl in val.split('\n')) + '\n'
 
 
-def indent(val: str, spaces=4) -> str:
+def indent(val: str, spaces: int = 4) -> str:
     if val == '':
         return val
     return ' ' * spaces + val.replace('\n', '\n' + ' ' * spaces)
@@ -46,9 +46,9 @@ def remove_indentation(source: str) -> str:
     for line in lines:
         if line and not line.isspace():
             indent_match = pattern.search(line)
-            if indent_match is not None:  # this is just for you mypy
+            if indent_match is not None:  # re.search() always returns a match here, never None
                 spaces.append(len(indent_match[0]))
 
     indent = min(spaces)
-    lines = [l[indent:] for l in lines]
+    lines = [line[indent:] for line in lines]
     return '\n'.join(lines)